+++ src/Components/Header/Header.js
@@ -1,9 +1,30 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
+import { useAuth } from '../Authorization/authContext';
 import './Header.css';
 
-//This file makes the header bar at the top of the screen and it adjusts depending on the size of the screen
+
+//This file makes the header bar at the top of the screen 
 const Header = () => {
+  const { user, logout } = useAuth(); 
+  const navigate = useNavigate();
+
+  const handleLogout = async () => {
+    try {
+      await logout(); // Ensure logout function is correctly imported and used
+    } catch (error) {
+      console.error('Error logging out', error);
+    }
+  };
+
+  const handleSurveyClick = () => {
+    if (!user) {
+      alert('You must log in to access the Survey.'); // Alongside the protected route, there is an alert notifying the user
+    } else {
+      navigate('/survey'); // Navigate to the survey route if user is logged in
+    }
+  };
+
   return (
     <div>
       {/* The large_only allows for items to vanish from nav bar when screen size shrinks
@@ -11,15 +32,26 @@ const Header = () => {
       <ul className="top_nav_bar">
         <li className="left"><Link to="/">Home</Link></li>
         <li className="right"><Link to="/about">About</Link></li>
-        <li className='right'><Link to="/survey">Survey</Link></li>
+        <li className='right'><Link to="/survey" onClick={handleSurveyClick}>Survey</Link></li>
         {/*<li class="right large_only"><a href="">Option 1</a></li>
         <li class="right large_only"><a href="">Option 2</a></li>
         <li class="right large_only"><a href="">Option 3</a></li>*/}
-        <li className="right"><Link to="/login">Log In</Link></li>
-        <li className="right"><Link to="/sign up">Sign Up</Link></li>
+        <li className="right"><Link to="/signup">Sign Up</Link></li>
+        {user ? (
+            <>
+              <li className='right'>
+                <button className='logoutButton' onClick={handleLogout}>Logout</button>
+              </li>
+            </>
+          ) : (
+        <li className="right"><Link to="/login">Login</Link></li>
+          )}
       </ul>
     </div>
   );
 };
 
 export default Header;
+
+
+