+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 # Changelog
+## [0.3.0] - 07-10-2024 - feature 5
+### Added (JP)
+- Created `ProtectedRoute.js` which protects the use of the full Survey 
+- Coded `login.js` and `signup.js` for authorization 
 
-## [0.1.5] - 07-01-2024 - feature 4 
+### Added (Grant)
+- 
+
+### Fixes
+-
+
+### Issues
+-
+
+### Progress Report
+-
+
+## [0.2.0] - 07-01-2024 - feature 4 
 ### Added (Grant)
 - Created `SongModel.js` and `GenreModel.js` for separate Parse queries.
 - Updated `Main.js` get async data for songs, genres, and users.
@@ -12,13 +28,13 @@
 - Routing for Home and Header are added as well as `Navigation.js` which is intended for routing
     - this routing will later be incorporated into the header bar as well as have its own page to function on
 
-### Fixed
+### Fixes
 - no longer calling for non existent css file in `index.html`.
 
 ### Current Problems/Issues
 - Genres and Songs won't connect through routing
 
-### Things in Progress
+### Progress Reports
 - About, Login and Sign Up are in the works for the next feature
 
 ## [0.1.0] - 06-20-2024 - feature 3